+++ src/components/tabs/script.ts
@@ -1,0 +1,21 @@
+import {
+  Component,
+  Prop,
+  Vue,
+} from 'vue-property-decorator';
+
+
+interface Tab {
+  label: string;
+  icon?: string;
+  closeable?: boolean;
+  active?: boolean;
+}
+
+@Component({
+  name: 'StTabs',
+})
+export default class StTabs extends Vue {
+  @Prop(Array)
+  tabs!: Tab[];
+}